feat: Add upgrade tests for TrustyAIService
