+++ tests/model_explainability/trustyai_service/trustyai_service_utils.py
@@ -1,3 +1,5 @@
+from __future__ import annotations
+
 import json
 import os
 from http import HTTPStatus
