+++ tests/model_explainability/trustyai_service/upgrade/test_trustyai_service_upgrade.py
@@ -4,10 +4,9 @@ from tests.model_explainability.trustyai_service.constants import DRIFT_BASE_DAT
 from tests.model_explainability.trustyai_service.trustyai_service_utils import (
     send_inferences_and_verify_trustyai_service_registered,
     verify_upload_data_to_trustyai_service,
-    verify_trustyai_service_metric_request,
+    verify_trustyai_service_metric_delete_request,
     TrustyAIServiceMetrics,
     verify_trustyai_service_metric_scheduling_request,
-    verify_trustyai_service_metric_delete_request,
 )
 from utilities.constants import MinIo
 from utilities.manifests.openvino import OPENVINO_KSERVE_INFERENCE_CONFIG
@@ -17,7 +16,7 @@ from utilities.manifests.openvino import OPENVINO_KSERVE_INFERENCE_CONFIG
     "model_namespace, minio_pod, minio_data_connection",
     [
         pytest.param(
-            {"name": "test-drift"},
+            {"name": "test-trustyaiservice-upgrade"},
             MinIo.PodConfig.MODEL_MESH_MINIO_CONFIG,
             {"bucket": MinIo.Buckets.MODELMESH_EXAMPLE_MODELS},
         )
@@ -25,18 +24,9 @@ from utilities.manifests.openvino import OPENVINO_KSERVE_INFERENCE_CONFIG
     indirect=True,
 )
 @pytest.mark.usefixtures("minio_pod")
-@pytest.mark.smoke
-class TestDriftMetrics:
-    """
-    Verifies all the basic operations with a drift metric (meanshift) available in TrustyAI, using PVC storage.
-
-    1. Send data to the model (gaussian_credit_model) and verify that TrustyAI registers the observations.
-    2. Send metric request (meanshift) and verify the response.
-    3. Send metric scheduling request and verify the response.
-    4. Send metric deletion request and verify that the scheduled metric has been deleted.
-    """
-
-    def test_drift_send_inference_and_verify_trustyai_service(
+class TestPreUpgradeTrustyAIService:
+    @pytest.mark.pre_upgrade
+    def test_trustyai_service_pre_upgrade_inference(
         self,
         admin_client,
         current_client_token,
@@ -44,6 +34,7 @@ class TestDriftMetrics:
         trustyai_service_with_pvc_storage,
         gaussian_credit_model,
     ) -> None:
+        """Set up a TrustyAIService with a model and inference before upgrade."""
         send_inferences_and_verify_trustyai_service_registered(
             client=admin_client,
             token=current_client_token,
@@ -53,13 +44,15 @@ class TestDriftMetrics:
             inference_config=OPENVINO_KSERVE_INFERENCE_CONFIG,
         )
 
-    def test_upload_data_to_trustyai_service(
+    @pytest.mark.pre_upgrade
+    def test_trustyai_service_pre_upgrade_data_upload(
         self,
         admin_client,
         minio_data_connection,
         current_client_token,
         trustyai_service_with_pvc_storage,
     ) -> None:
+        """"""
         verify_upload_data_to_trustyai_service(
             client=admin_client,
             trustyai_service=trustyai_service_with_pvc_storage,
@@ -67,24 +60,7 @@ class TestDriftMetrics:
             data_path=f"{DRIFT_BASE_DATA_PATH}/training_data.json",
         )
 
-    def test_drift_metric_meanshift(
-        self,
-        admin_client,
-        current_client_token,
-        trustyai_service_with_pvc_storage,
-        gaussian_credit_model,
-    ):
-        verify_trustyai_service_metric_request(
-            client=admin_client,
-            trustyai_service=trustyai_service_with_pvc_storage,
-            token=current_client_token,
-            metric_name=TrustyAIServiceMetrics.Drift.MEANSHIFT,
-            json_data={
-                "modelId": gaussian_credit_model.name,
-                "referenceTag": "TRAINING",
-            },
-        )
-
+    @pytest.mark.pre_upgrade
     def test_drift_metric_schedule_meanshift(
         self,
         admin_client,
@@ -103,6 +79,21 @@ class TestDriftMetrics:
             },
         )
 
+
+@pytest.mark.parametrize(
+    "model_namespace, minio_pod, minio_data_connection",
+    [
+        pytest.param(
+            {"name": "test-trustyaiservice-upgrade"},
+            MinIo.PodConfig.MODEL_MESH_MINIO_CONFIG,
+            {"bucket": MinIo.Buckets.MODELMESH_EXAMPLE_MODELS},
+        )
+    ],
+    indirect=True,
+)
+@pytest.mark.usefixtures("minio_pod")
+class TestPostUpgradeTrustyAIService:
+    @pytest.mark.post_upgrade
     def test_drift_metric_delete(
         self,
         admin_client,
